backend/requirements.txt
amqp==5.3.1
astroid==3.3.8
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dill==0.3.9
dnspython==1.16.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
humanize==4.12.1
idna==3.10
iniconfig==2.0.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
msgspec==0.19.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pycparser==2.22
pylint==3.3.4
pymongo==3.12.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
rdkit==2024.9.5
redis==5.2.1
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
tomlkit==0.13.2
tornado==6.4.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
paho-mqtt==2.1.0
